gen1/requirements.txt
#functions-framework==3.3.0
google-cloud-storage==2.9.0
google-cloud-logging==3.5.0
redis==4.5.5
